+++ src/core/markdown/generate-docs.ts
@@ -21,17 +21,17 @@ import {
   UnparsedSourceBundle,
 } from '../shared/types';
 import { parsedFilesToRenderableBundle } from './adapters/renderable-bundle';
-import { reflectApexSource } from '../reflection/reflect-source';
-import { addInheritanceChainToTypes } from '../reflection/inheritance-chain-expanion';
-import { addInheritedMembersToTypes } from '../reflection/inherited-member-expansion';
+import { reflectApexSource } from '../reflection/apex/reflect-apex-source';
+import { addInheritanceChainToTypes } from '../reflection/apex/inheritance-chain-expanion';
+import { addInheritedMembersToTypes } from '../reflection/apex/inherited-member-expansion';
 import { convertToDocumentationBundle } from './adapters/renderable-to-page-data';
-import { filterScope } from '../reflection/filter-scope';
+import { filterScope } from '../reflection/apex/filter-scope';
 import { Template } from '../template';
 import { hookableTemplate } from './templates/hookable';
 import { sortTypesAndMembers } from '../reflection/sort-types-and-members';
 import { isSkip } from '../shared/utils';
 import { parsedFilesToReferenceGuide } from './adapters/reference-guide';
-import { removeExcludedTags } from '../reflection/remove-excluded-tags';
+import { removeExcludedTags } from '../reflection/apex/remove-excluded-tags';
 import { HookError } from '../errors/errors';
 import { reflectObjectSources } from '../reflection/reflect-object-source';
