+++ src/core/reflection/apex/remove-excluded-tags.ts
@@ -3,7 +3,7 @@ import { match } from 'fp-ts/boolean';
 import { ClassMirror, DocComment, InterfaceMirror, Type } from '@cparra/apex-reflection';
 import { pipe } from 'fp-ts/function';
 import { apply } from '#utils/fp';
-import { ParsedFile } from '../shared/types';
+import { ParsedFile } from '../../shared/types';
 
 type AppliedRemoveTagFn = (tagName: string, removeFn: RemoveTagFn) => DocComment;
 type RemoveTagFn = (docComment: DocComment) => DocComment;
